Move the "extras" data to be in their own profile netCDF file
This assumes that the "extras" data occurs before or after a profile
and for now I think that is OK. This reduces the complexity of the
code quite a bit. We just make the "extras" data look like a profile
dataset, optionally setting the depths to zero if there are no depths
to capture the data as surface measurements.

This also changes some of the extras configuration kwarg names so we
can control the creation of the ASCII pseudogram and the inclustion
of that data into the netCDF profile files, and the image creation.